Tests/LibWeb/Text/expected/hit_testing/pointer-events.txt

<DIV id="a1">
<BODY>
---
<DIV id="b3">
<DIV id="b2">
---
<A id="c1">
<BODY>
---
<I id="d2">
<B id="d1">
---
<DIV id="e2">
<BODY>
---
